# Flatten `Status.tagged_with_all` scope