@renchap 2024-03-27
If you have sample SQL queries (and Rails commands to find relevant objects to run them with), we can try running them on mastodon.social