@mjankowski 2024-03-27
> If you have sample SQL queries (and Rails commands to find relevant objects to run them with), we can try running them on mastodon.social and see how is the perf

Good idea - will add something shortly. I just realized that the change here fails to handle the "nothing passed in" case. Going to add unit spec for that, and update. After that, will post some benchmark ideas.