@mjankowski 2024-03-27
Updated with a commit to just avoiding calling the scope w/ nil values for now.

For benchmarking ... you could grab some collection of the most/least used tag ids as an array in irb, and then pass them through the original and changed method. For my local run I renamed the old one `tagged_with_all_original` and ran them side by side a bunch. You could call the new one something else, or just using the scope chain directly w/out defining a scope/method.

May be useful to put an `explain` after each to see the full plan cascade effect as well.