@mjankowski 2024-10-23
Did we ever do any benchmarking on this, or do I need to provide a more detailed code snippet to use?